Write only changed (#1153)
* Only write changed files

* Only copy item if file contents changed (#1152)

* Only copy item if file contents changed

* Fix linting issue


* Add --writeOnlyChanged

* Fix VirtualProject.ts

* Fix createProjectData.ts

* Fix init.ts

+++ src/Project/functions/compileFiles.ts
@@ -140,7 +140,13 @@ export function compileFiles(
 		benchmarkIfVerbose("writing compiled files", () => {
 			for (const { sourceFile, source } of fileWriteQueue) {
 				const outPath = services.pathTranslator.getOutputPath(sourceFile.fileName);
-				fs.outputFileSync(outPath, source);
+				if (
+					!data.writeOnlyChanged ||
+					!fs.pathExistsSync(outPath) ||
+					fs.readFileSync(outPath).toString() !== source
+				) {
+					fs.outputFileSync(outPath, source);
+				}
 				if (compilerOptions.declaration) {
 					program.emit(sourceFile, ts.sys.writeFile, undefined, true, {
 						afterDeclarations: [transformTypeReferenceDirectives, transformPaths],
